+++ src/Compze.Typermedia.Hosting/ITypermediaTransportServer.cs
@@ -1,9 +1,0 @@
-namespace Compze.Typermedia.Hosting;
-
-public interface ITypermediaTransportServer : IAsyncDisposable
-{
-   Uri Address { get; }
-
-   Task StartAsync();
-   Task StopAsync();
-}